requirements.txt
autopep8==1.4.4
beautifulsoup4==4.8.1
colorama==0.4.1
Django==2.2.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-form==3.4
django-bootstrap4==1.0.1
django-flatpickr==1.0.1
future==0.17.1
importlib-metadata==0.23
isort==4.3.21
jedi==0.14.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
mysql-connector==2.2.9
parso==0.5.1
Pillow==6.2.0
pluggy==0.13.0
pycodestyle==2.5.0
pydocstyle==4.0.1
pyflakes==2.1.1
pylint==2.4.2
pyls==0.1.6
python-jsonrpc-server==0.2.0
python-language-server==0.28.3
pytz==2019.3
rope==0.14.0
six==1.12.0
snowballstemmer==2.0.0
soupsieve==1.9.4
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
yapf==0.28.0
zipp==0.6.0
